Face/requirements.txt
# Face Recognition and Embedding Extraction Dependencies

# Core dependencies
opencv-python>=4.8.0
dlib>=19.24.0
deepface>=0.0.87
numpy>=1.21.0
imutils>=0.5.4
scipy>=1.9.0
matplotlib>=3.5.0
tensorflow>=2.10.0
Pillow>=9.0.0
cmake>=3.20.0
scikit-learn>=1.0.0
PyYAML>=6.0

# Additional dependencies for face embedding extraction
retina-face>=0.0.13
mtcnn>=0.1.1
face-recognition>=1.3.0
